Add files via upload
Guess problem: In one family 10 members are guessing a number, for example we have the actual number input but none of the members knows this number, they will have one chance to guess, so 10 inputs from 10 members and then person who guessed the closest will be the winner in the list. Write a simple function for this problem, where actual number and user inputs are passed.